+++ tests/Console/ModelsCommand/Relations/__snapshots__/Test__test__1.php
@@ -5,6 +5,7 @@ declare(strict_types=1);
 namespace Barryvdh\LaravelIdeHelper\Tests\Console\ModelsCommand\Relations\Models;
 
 use Barryvdh\LaravelIdeHelper\Tests\Console\ModelsCommand\Relations\ModelsOtherNamespace\AnotherModel;
+use Barryvdh\LaravelIdeHelper\Tests\Console\ModelsCommand\Relations\Traits\HasTestRelations;
 use Illuminate\Database\Eloquent\Model;
 use Illuminate\Database\Eloquent\Relations\BelongsTo;
 use Illuminate\Database\Eloquent\Relations\BelongsToMany;
@@ -38,6 +39,9 @@ use Illuminate\Database\Eloquent\Relations\MorphToMany;
  * @property-read Model|\Eloquent $relationMorphTo
  * @property-read \Illuminate\Database\Eloquent\Collection|Simple[] $relationMorphedByMany
  * @property-read int|null $relation_morphed_by_many_count
+ * @property-read \Illuminate\Database\Eloquent\Collection|Simple[] $relationSampleRelationType
+ * @property-read int|null $relation_sample_relation_type_count
+ * @property-read Simple $relationSampleToManyRelationType
  * @method static \Illuminate\Database\Eloquent\Builder|Simple newModelQuery()
  * @method static \Illuminate\Database\Eloquent\Builder|Simple newQuery()
  * @method static \Illuminate\Database\Eloquent\Builder|Simple query()
@@ -46,6 +50,8 @@ use Illuminate\Database\Eloquent\Relations\MorphToMany;
  */
 class Simple extends Model
 {
+    use HasTestRelations;
+
     // Regular relations
     public function relationHasMany(): HasMany
     {
@@ -112,5 +118,15 @@ class Simple extends Model
     public function relationBelongsToSameNameAsColumn(): BelongsTo
     {
         return $this->belongsTo(AnotherModel::class, __FUNCTION__);
+    }
+
+    public function relationSampleToManyRelationType()
+    {
+        return $this->testToOneRelation(Simple::class);
+    }
+
+    public function relationSampleRelationType()
+    {
+        return $this->testToManyRelation(Simple::class);
     }
 }